feat(Users): Add `users find` CLI command
